// src/module/issue/enum/message.enum.ts
export enum Message {
  NoIssue = 'The issue does not exist',
  SuccessUpdate = 'The issue has been successfully updated!',
  SuccessCreate = 'The issue has been successfully created!',
  SuccessDelete = 'The issue has been successfully deleted!',

  RequireTitle = 'Title is required',
  RequireDescription = 'Description is required',
}